% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqToolsSRJ-package.R
\docType{package}
\name{SeqToolsSRJ}
\alias{SeqToolsSRJ}
\alias{SeqToolsSRJ-package}
\title{SeqToolsSRJ - Support for doing sequencing informatics}
\description{
This is a dollection of functions I use to do sequencing
informatics.
}
